# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalTargetResponse": {
        "additionalProperties": false,
        "properties": {
          "approved": {
            "title": "Approved",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "name",
          "approved"
        ],
        "title": "ApprovalTargetResponse",
        "type": "object"
      },
      "AuthorResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "AuthorResponse",
        "type": "object"
      },
      "BoardResponse": {
        "additionalProperties": false,
        "properties": {
          "editing": {
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "title": "Editing",
            "type": "array"
          },
          "final": {
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "title": "Final",
            "type": "array"
          },
          "review": {
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "title": "Review",
            "type": "array"
          },
          "selection": {
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "title": "Selection",
            "type": "array"
          }
        },
        "required": [
          "selection",
          "editing",
          "review",
          "final"
        ],
        "title": "BoardResponse",
        "type": "object"
      },
...
      "CommentCreate": {
        "properties": {
          "body": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          },
          "kind": {
            "default": "comment",
            "enum": [
              "comment",
              "change_request"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "CommentCreate",
        "type": "object"
      },
      "CommentResponse": {
        "additionalProperties": false,
        "properties": {
          "author": {
            "$ref": "#/components/schemas/AuthorResponse"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "author",
          "body",
          "kind",
          "created_at"
        ],
        "title": "CommentResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "ItemsResponse_NotificationResponse_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotificationResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "ItemsResponse[NotificationResponse]",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "NotificationResponse": {
        "additionalProperties": false,
        "properties": {
          "album_id": {
            "title": "Album Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "photo_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Id"
          },
          "read": {
            "title": "Read",
            "type": "boolean"
          },
          "version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Id"
          }
        },
        "required": [
          "id",
          "album_id",
          "photo_id",
          "version_id",
          "kind",
          "message",
          "read",
          "created_at"
        ],
        "title": "NotificationResponse",
        "type": "object"
      },
      "OkResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
...
      "PersonResponse": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "link_status": {
            "title": "Link Status",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "proposed_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed User Id"
          },
          "reference_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Url"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "name",
          "user_id",
          "proposed_user_id",
          "reference_url",
          "link_status"
        ],
        "title": "PersonResponse",
        "type": "object"
      },
...
      "PhotoResponse": {
        "additionalProperties": false,
        "properties": {
          "album_id": {
            "title": "Album Id",
            "type": "string"
          },
          "analysis_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Error"
          },
          "analysis_metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Analysis Metadata",
            "type": "object"
          },
          "analysis_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Mode"
          },
          "analysis_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Provider"
          },
          "analysis_status": {
            "title": "Analysis Status",
            "type": "string"
          },
          "board_status": {
            "title": "Board Status",
            "type": "string"
          },
          "capture_timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capture Timezone"
          },
          "captured_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Captured At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_url": {
            "title": "Display Url",
            "type": "string"
          },
          "face_count": {
            "title": "Face Count",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "final_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Version Id"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Name"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "original_url": {
            "title": "Original Url",
            "type": "string"
          },
          "people": {
            "items": {
              "$ref": "#/components/schemas/PersonResponse"
            },
            "title": "People",
            "type": "array"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "quality": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Quality",
            "type": "object"
          },
          "selected": {
            "title": "Selected",
            "type": "boolean"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "thumbnail_url": {
            "title": "Thumbnail Url",
            "type": "string"
          },
          "unknown_faces": {
            "title": "Unknown Faces",
            "type": "integer"
          },
          "uploader_id": {
            "title": "Uploader Id",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "album_id",
          "uploader_id",
          "filename",
          "thumbnail_url",
          "display_url",
          "original_url",
          "width",
          "height",
          "created_at",
          "captured_at",
          "capture_timezone",
          "latitude",
          "longitude",
          "location_name",
          "analysis_status",
          "analysis_error",
          "analysis_provider",
          "analysis_mode",
          "analysis_metadata",
          "face_count",
          "unknown_faces",
          "people",
          "tags",
          "purpose",
          "selected",
          "note",
          "final_version_id",
          "board_status",
          "quality"
        ],
        "title": "PhotoResponse",
        "type": "object"
      },
...
      "VersionResponse": {
        "additionalProperties": false,
        "properties": {
          "approval_count": {
            "title": "Approval Count",
            "type": "integer"
          },
          "author": {
            "$ref": "#/components/schemas/AuthorResponse"
          },
          "brightness": {
            "title": "Brightness",
            "type": "number"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/CommentResponse"
            },
            "title": "Comments",
            "type": "array"
          },
          "consensus": {
            "title": "Consensus",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_final": {
            "title": "Is Final",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "photo_id": {
            "title": "Photo Id",
            "type": "string"
          },
          "preview_url": {
            "title": "Preview Url",
            "type": "string"
          },
          "renderer_version": {
            "title": "Renderer Version",
            "type": "string"
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason"
          },
          "review_requested": {
            "title": "Review Requested",
            "type": "boolean"
          },
          "saturation": {
            "title": "Saturation",
            "type": "number"
          },
          "target_count": {
            "title": "Target Count",
            "type": "integer"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/ApprovalTargetResponse"
            },
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "id",
          "photo_id",
          "number",
          "name",
          "parent_id",
          "author",
          "created_at",
          "brightness",
          "saturation",
          "renderer_version",
          "preview_url",
          "review_requested",
          "needs_review",
          "review_reason",
          "targets",
          "approval_count",
          "target_count",
          "consensus",
          "is_final",
          "comments"
        ],
        "title": "VersionResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ZZIK API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/albums/{album_id}/board": {
      "get": {
        "operationId": "board_api_albums__album_id__board_get",
        "parameters": [
          {
            "in": "path",
            "name": "album_id",
            "required": true,
            "schema": {
              "title": "Album Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BoardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Board"
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "notifications_api_notifications_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemsResponse_NotificationResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Notifications"
      }
    },
    "/api/notifications/{notification_id}/read": {
      "post": {
        "operationId": "read_notification_api_notifications__notification_id__read_post",
        "parameters": [
          {
            "in": "path",
            "name": "notification_id",
            "required": true,
            "schema": {
              "title": "Notification Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Read Notification"
      }
    },
...
    "/api/versions/{version_id}/comments": {
      "post": {
        "operationId": "add_comment_api_versions__version_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Content Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Add Comment"
      }
    },
...
  }
}